[cleanup] move qphix interface to dedicated directory
