+++ qphix/qphix_veclen.h
